[cache-machine] minor change in context.py
