+++ pyblazing/pyblazing/apiv2/context.py
@@ -303,7 +303,77 @@ def mergeMetadataFor(curr_table, fileMetadata, hiveMetadata, extra_columns):
     frame = OrderedDict(((key,value) for (key,value) in zip(names, series)))
     result = cudf.DataFrame(frame)
     return result
-    
+
+
+import json
+import collections
+
+def visit (lines): 
+	stack = collections.deque()
+	root_level = 0
+	dicc = {
+		"expr": lines[root_level][1],
+		"children": []
+	}
+	processed = set()
+	for index in range(len(lines)): 
+		child_level, expr = lines[index]
+		if child_level == root_level + 1:
+			new_dicc = {
+				"expr": expr,
+				"children": []
+			}
+			if len(dicc["children"]) == 0:
+				dicc["children"] = [new_dicc]
+			else:
+				dicc["children"].append(new_dicc)
+			stack.append( (index, child_level, expr, new_dicc) )
+			processed.add(index)
+
+	for index in processed:
+		lines[index][0] = -1 
+
+	while len(stack) > 0: 
+		curr_index, curr_level, curr_expr, curr_dicc = stack.pop()
+		processed = set()
+
+		if curr_index < len(lines)-1: # is brother
+			child_level, expr = lines[curr_index+1]
+			if child_level == curr_level:
+				continue
+			elif child_level == curr_level + 1:
+				index = curr_index + 1				
+				while index < len(lines) and child_level == curr_level + 1: 
+					child_level, expr = lines[index]
+					if child_level != curr_level + 1:
+						continue
+					first = index
+					new_dicc = {
+						"expr": expr,
+						"children": []
+					}
+					if len(curr_dicc["children"]) == 0:
+						curr_dicc["children"] = [new_dicc]
+					else:
+						curr_dicc["children"].append(new_dicc)
+					processed.add(index)
+					stack.append( (index, child_level, expr, new_dicc) )
+					index += 1
+		for index in processed:
+			lines[index][0] = -1 
+	return json.dumps(dicc)
+
+
+def get_plan(algebra):
+	algebra = algebra.replace("  ", "\t")
+	lines = algebra.split("\n")
+	new_lines = [] 
+	for i in range(len(lines) - 1):
+		line = lines[i]
+		level = line.count("\t")
+		new_lines.append( [level, line.replace("\t", "")] )
+
+	return visit(new_lines)    
 
 class BlazingTable(object):
     def __init__(
@@ -868,7 +938,7 @@ class BlazingContext(object):
         return result
 
 
-    def execute(self, sql, algebra, plan):
+    def execute(self, sql, algebra):
         # TODO: remove hardcoding
         masterIndex = 0
         nodeTableList = [{} for _ in range(len(self.nodes))]
@@ -888,7 +958,7 @@ class BlazingContext(object):
                 workers=[worker])
             new_tables, relational_algebra_steps = connection.result()
 
-        # algebra = modifyAlgebraForDataframesWithOnlyWantedColumns(algebra, relational_algebra_steps,self.tables)
+        algebra = modifyAlgebraForDataframesWithOnlyWantedColumns(algebra, relational_algebra_steps,self.tables)
 
         for table in new_tables:
             fileTypes.append(new_tables[table].fileType)
@@ -911,7 +981,7 @@ class BlazingContext(object):
 
         ctxToken = random.randint(0, 64000)
         accessToken = 0
-
+        json_plan = get_plan(algebra)
         if self.dask_client is None:
             result = cio.runQueryCaller(
                         masterIndex,
@@ -919,7 +989,7 @@ class BlazingContext(object):
                         nodeTableList[0],
                         fileTypes,
                         ctxToken,
-                        plan,
+                        json_plan,
                         accessToken)
         else:
             dask_futures = []
@@ -934,7 +1004,7 @@ class BlazingContext(object):
                         nodeTableList[i],
                         fileTypes,
                         ctxToken,
-                        plan,
+                        json_plan,
                         accessToken,
                         workers=[worker]))
                 i = i + 1
